app/requirements.txt
Flask==2.1.1
Flask-MySQL==1.5.1
mysql-connector-python==8.0.29
